# Speed up GcsIO throughput by 10x