@charlesccychen 2016-11-17
Thanks for the quick review.

R: @robertwb for merge.