@charlesccychen 2016-11-17
Note that the Jenkins failure is due to Java and therefore spurious.  My personal Travis CI passes: https://travis-ci.org/charlesccychen/incubator-beam/builds/176809464